+++ core/src/config/ServerConfig.h
@@ -92,15 +92,9 @@ struct ServerConfig {
         } http;
     } network;
 
-    struct DB {
-        Floating archive_disk_threshold{0.0};
-        Integer archive_days_threshold{0};
-    } db;
-
     struct Storage {
         String path{"unknown"};
         Integer auto_flush_interval{0};
-        Integer file_cleanup_timeout{0};
     } storage;
 
     struct Cache {
@@ -139,7 +133,6 @@ struct ServerConfig {
 
     struct WAL {
         Bool enable{false};
-        Bool recovery_error_ignore{false};
         Integer buffer_size{0};
         String path{"unknown"};
     } wal;
