+++ core/src/db/DBFactory.h
@@ -12,7 +12,6 @@
 #pragma once
 
 #include "DB.h"
-#include "Options.h"
 
 #include <memory>
 #include <string>
