+++ core/src/db/Types.cpp
@@ -30,6 +30,7 @@ const char* PARAM_INDEX_EXTRA_PARAMS = "params";
 const char* PARAM_SEGMENT_ROW_COUNT = "segment_row_count";
 
 const char* DEFAULT_STRUCTURED_INDEX = "SORTED";  // this string should be defined in knowhere::IndexEnum
+const char* DEFAULT_PARTITON_TAG = "_default";
 
 }  // namespace engine
 }  // namespace milvus
