+++ core/src/db/Types.h
@@ -30,6 +30,24 @@
 namespace milvus {
 namespace engine {
 
+extern const char* FIELD_UID;
+
+extern const char* ELEMENT_RAW_DATA;
+extern const char* ELEMENT_BLOOM_FILTER;
+extern const char* ELEMENT_DELETED_DOCS;
+extern const char* ELEMENT_INDEX_COMPRESS;
+
+extern const char* PARAM_UID_AUTOGEN;
+extern const char* PARAM_DIMENSION;
+extern const char* PARAM_INDEX_TYPE;
+extern const char* PARAM_INDEX_METRIC_TYPE;
+extern const char* PARAM_INDEX_EXTRA_PARAMS;
+extern const char* PARAM_SEGMENT_ROW_COUNT;
+
+extern const char* DEFAULT_STRUCTURED_INDEX;
+extern const char* DEFAULT_PARTITON_TAG;
+
+///////////////////////////////////////////////////////////////////////////////////////////////////
 using id_t = int64_t;
 using offset_t = int32_t;
 using date_t = int32_t;
@@ -42,6 +60,7 @@ using VectorDistances = std::vector<VectorDistance>;
 using ResultIds = std::vector<faiss::Index::idx_t>;
 using ResultDistances = std::vector<faiss::Index::distance_t>;
 
+///////////////////////////////////////////////////////////////////////////////////////////////////
 enum DataType {
     NONE = 0,
     BOOL = 1,
@@ -59,6 +78,17 @@ enum DataType {
     VECTOR_FLOAT = 101,
 };
 
+///////////////////////////////////////////////////////////////////////////////////////////////////
+enum FieldElementType {
+    FET_NONE = 0,
+    FET_RAW = 1,
+    FET_BLOOM_FILTER = 2,
+    FET_DELETED_DOCS = 3,
+    FET_INDEX = 4,
+    FET_COMPRESS_SQ8 = 5,
+};
+
+///////////////////////////////////////////////////////////////////////////////////////////////////
 class BinaryData : public cache::DataObj {
  public:
     int64_t
@@ -71,6 +101,7 @@ class BinaryData : public cache::DataObj {
 };
 using BinaryDataPtr = std::shared_ptr<BinaryData>;
 
+///////////////////////////////////////////////////////////////////////////////////////////////////
 class VaribleData : public cache::DataObj {
  public:
     int64_t
@@ -84,6 +115,7 @@ class VaribleData : public cache::DataObj {
 };
 using VaribleDataPtr = std::shared_ptr<VaribleData>;
 
+///////////////////////////////////////////////////////////////////////////////////////////////////
 using FIELD_TYPE_MAP = std::unordered_map<std::string, DataType>;
 using FIELD_WIDTH_MAP = std::unordered_map<std::string, int64_t>;
 using FIXEDX_FIELD_MAP = std::unordered_map<std::string, BinaryDataPtr>;
@@ -91,14 +123,15 @@ using VARIABLE_FIELD_MAP = std::unordered_map<std::string, VaribleDataPtr>;
 using VECTOR_INDEX_MAP = std::unordered_map<std::string, knowhere::VecIndexPtr>;
 using STRUCTURED_INDEX_MAP = std::unordered_map<std::string, knowhere::IndexPtr>;
 
+///////////////////////////////////////////////////////////////////////////////////////////////////
 struct DataChunk {
     int64_t count_ = 0;
     FIXEDX_FIELD_MAP fixed_fields_;
     VARIABLE_FIELD_MAP variable_fields_;
 };
-
 using DataChunkPtr = std::shared_ptr<DataChunk>;
 
+///////////////////////////////////////////////////////////////////////////////////////////////////
 struct CollectionIndex {
     std::string index_name_;
     std::string index_type_;
@@ -106,6 +139,7 @@ struct CollectionIndex {
     milvus::json extra_params_ = {{"nlist", 2048}};
 };
 
+///////////////////////////////////////////////////////////////////////////////////////////////////
 struct VectorsData {
     uint64_t vector_count_ = 0;
     std::vector<float> float_data_;
@@ -113,13 +147,7 @@ struct VectorsData {
     IDNumbers id_array_;
 };
 
-struct Entity {
-    int64_t entity_count_ = 0;
-    std::vector<uint8_t> attr_value_;
-    std::unordered_map<std::string, VectorsData> vector_data_;
-    IDNumbers id_array_;
-};
-
+///////////////////////////////////////////////////////////////////////////////////////////////////
 struct AttrsData {
     uint64_t attr_count_ = 0;
     std::unordered_map<std::string, engine::DataType> attr_type_;
@@ -127,6 +155,7 @@ struct AttrsData {
     IDNumbers id_array_;
 };
 
+///////////////////////////////////////////////////////////////////////////////////////////////////
 struct QueryResult {
     uint64_t row_num_;
     engine::ResultIds result_ids_;
@@ -135,33 +164,31 @@ struct QueryResult {
 };
 using QueryResultPtr = std::shared_ptr<QueryResult>;
 
-using File2ErrArray = std::map<std::string, std::vector<std::string>>;
-using Table2FileErr = std::map<std::string, File2ErrArray>;
+///////////////////////////////////////////////////////////////////////////////////////////////////
+struct DBMetaOptions {
+    std::string path_;
+    std::string backend_uri_;
+};  // DBMetaOptions
 
-extern const char* FIELD_UID;
+///////////////////////////////////////////////////////////////////////////////////////////////////
+struct DBOptions {
+    typedef enum { SINGLE = 0, CLUSTER_READONLY, CLUSTER_WRITABLE } MODE;
 
-extern const char* ELEMENT_RAW_DATA;
-extern const char* ELEMENT_BLOOM_FILTER;
-extern const char* ELEMENT_DELETED_DOCS;
-extern const char* ELEMENT_INDEX_COMPRESS;
+    DBMetaOptions meta_;
+    int mode_ = MODE::SINGLE;
 
-extern const char* PARAM_UID_AUTOGEN;
-extern const char* PARAM_DIMENSION;
-extern const char* PARAM_INDEX_TYPE;
-extern const char* PARAM_INDEX_METRIC_TYPE;
-extern const char* PARAM_INDEX_EXTRA_PARAMS;
-extern const char* PARAM_SEGMENT_ROW_COUNT;
+    size_t insert_buffer_size_ = 4 * GB;
+    bool insert_cache_immediately_ = false;
 
-extern const char* DEFAULT_STRUCTURED_INDEX;
+    int64_t auto_flush_interval_ = 1;
 
-enum FieldElementType {
-    FET_NONE = 0,
-    FET_RAW = 1,
-    FET_BLOOM_FILTER = 2,
-    FET_DELETED_DOCS = 3,
-    FET_INDEX = 4,
-    FET_COMPRESS_SQ8 = 5,
-};
+    bool metric_enable_ = false;
+
+    // wal relative configurations
+    bool wal_enable_ = true;
+    int64_t buffer_size_ = 256;
+    std::string mxlog_path_ = "/tmp/milvus/wal/";
+};  // Options
 
 }  // namespace engine
 }  // namespace milvus
