+++ core/src/db/insert/MemManagerFactory.h
@@ -12,7 +12,6 @@
 #pragma once
 
 #include "MemManager.h"
-#include "db/Options.h"
 
 #include <memory>
