+++ core/src/server/DBWrapper.cpp
@@ -41,7 +41,6 @@ DBWrapper::StartService() {
     opt.meta_.path_ = path + "/db";
 
     opt.auto_flush_interval_ = config.storage.auto_flush_interval();
-    opt.file_cleanup_timeout_ = config.storage.file_cleanup_timeout();
     opt.metric_enable_ = config.metric.enable();
     opt.insert_cache_immediately_ = config.cache.cache_insert_data();
     opt.insert_buffer_size_ = config.cache.insert_buffer_size();
@@ -63,7 +62,6 @@ DBWrapper::StartService() {
     opt.wal_enable_ = false;
 
     if (opt.wal_enable_) {
-        opt.recovery_error_ignore_ = config.wal.recovery_error_ignore();
         int64_t wal_buffer_size = config.wal.buffer_size();
         wal_buffer_size /= (1024 * 1024);
         opt.buffer_size_ = wal_buffer_size;
@@ -87,20 +85,6 @@ DBWrapper::StartService() {
     // init faiss global variable
     int64_t use_blas_threshold = config.engine.use_blas_threshold();
     faiss::distance_compute_blas_threshold = use_blas_threshold;
-
-    // set archive config
-    engine::ArchiveConf::CriteriaT criterial;
-    int64_t disk = config.db.archive_disk_threshold();
-    int64_t days = config.db.archive_days_threshold();
-
-    if (disk > 0) {
-        criterial[engine::ARCHIVE_CONF_DISK] = disk;
-    }
-
-    if (days > 0) {
-        criterial[engine::ARCHIVE_CONF_DAYS] = days;
-    }
-    opt.meta_.archive_conf_.SetCriterias(criterial);
 
     // create db root folder
     s = CommonUtil::CreateDirectory(opt.meta_.path_);
